📝 chore(MediaController.php): move MediaController to Starter namespace for better organization and consistency 📝 chore(Media.php): replace deprecated Illuminate\Support\Facades\Storage with Illuminate\Support\Facades\Route to fix route generation for media URL 📝 chore(InstallCommand.php): add missing Starter directory for Starter controllers to ensure proper directory structure 📝 chore(InitMediaCommand.php): move MediaController to Starter namespace for better organization and consistency 📝 chore(AccountController.php): add AccountController to Starter namespace for better organization and consistency 📝 chore(GlobalSearchController.php): add GlobalSearchController to Starter namespace for better organization and consistency 📝 chore(NotificationController.php): add NotificationController to Starter namespace for better organization and consistency 📝 chore(UserLogController.php): add UserLogController to Starter namespace for better organization and consistency 📝 chore(routes/starter.php): update namespace for Starter controllers to match new directory structure
