+++ stubs/default/app/Http/Controllers/Starter/GlobalSearchController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Starter;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Pkt\StarterKit\Helpers\GlobalSearch;
