+++ stubs/default/app/Http/Controllers/Starter/NotificationController.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Starter;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
-use Carbon\Carbon;
-use Illuminate\Pagination\LengthAwarePaginator;
 
 class NotificationController extends Controller
 {
