+++ stubs/default/app/Http/Controllers/Starter/UserLogController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Starter;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
